requirements.txt
Flask==2.3.2
opencv-python==4.8.0.76
ultralytics==8.0.111
pyserial==3.5
